vserver/pyproject.toml
[tool.poetry]
name = "vserver"
version = "0.1.0"
description = ""
authors = ["Huy TQ <5723282+imhuytq@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.19.0"
fastapi = "^0.86.0"
celery = {extras = ["redis"], version = "^5.2.7"}
yt-dlp = "^2022.10.4"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.1"
sqlalchemy = "^1.4.43"
pydantic = "^1.10.2"
psycopg2-binary = "^2.9.5"
alembic = "^1.8.1"
python-ulid = "^1.1.0"
stringcase = "^1.2.0"
inflect = "^6.0.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
isort = "^5.10.1"
mypy = "^0.982"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
